MLE/assignment-2-victorw271/requirements.txt
setuptools
pytest
openml~=0.13.0
scikit-learn~=1.2.0
tqdm~=4.64.1
jupyterlab~=3.6.1
matplotlib~=3.6.3
scikit-image~=0.19.3
seaborn~=0.12.2
ipywidgets~=8.0.4
geopandas~=0.12.2
contextily~= 1.3.0
dirty_cat~= 0.4.0
category_encoders~=2.6.0
gensim==3.8.3
